requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
tqdm>=4.66.4
python-dotenv>=1.0.1 